chore: refactor to support component handler registration
