+++ v3/src/components/calculator/calculator-defs.ts
@@ -1,2 +1,3 @@
 export const kCalculatorTileType = "Calculator"
+export const kV2CalculatorType = "calculator"
 export const kCalculatorTileClass = "calculator"
