+++ v3/src/components/case-card/case-card-defs.ts
@@ -1,2 +1,3 @@
 export const kCaseCardTileType = "CaseCard"
+export const kV2CaseCardType = "caseCard"
 export const kCaseCardTileClass = "codap-case-card"
