+++ v3/src/components/case-card/case-card-registration.ts
@@ -1,7 +1,9 @@
+import { caseTableCardComponentHandler } from "../case-table-card-common/case-table-card-handler"
+import { registerComponentHandler } from "../../data-interactive/handlers/component-handler"
 import { registerTileComponentInfo } from "../../models/tiles/tile-component-info"
 import { registerTileContentInfo } from "../../models/tiles/tile-content-info"
 import { CaseCardComponent } from "./case-card-component"
-import { kCaseCardTileType } from "./case-card-defs"
+import { kCaseCardTileType, kV2CaseCardType } from "./case-card-defs"
 import { CaseCardModel } from "./case-card-model"
 import { CaseTableCardTitleBar } from "../case-table-card-common/case-table-card-title-bar"
 import CardIcon from '../../assets/icons/icon-case-card.svg'
@@ -35,6 +37,8 @@ registerTileComponentInfo({
   defaultWidth: 200,
   defaultHeight: 400
 })
+
+registerComponentHandler(kV2CaseCardType, caseTableCardComponentHandler)
 
 /*
 registerV2TileImporter("DG.CardView", ({ v2Component, v2Document, sharedModelManager, insertTile }) => {
