+++ v3/src/components/case-table/case-table-defs.ts
@@ -1,2 +1,3 @@
 export const kCaseTableTileType = "CaseTable"
+export const kV2CaseTableType = "caseTable"
 export const kCaseTableTileClass = "codap-case-table"
