+++ v3/src/components/graph/graph-defs.ts
@@ -1,4 +1,5 @@
 export const kGraphTileType = "Graph"
+export const kV2GraphType = "graph"
 export const kGraphTileClass = "codap-graph"
 
 export const kGraphIdPrefix = "GRPH"
