+++ v3/src/components/map/map-defs.ts
@@ -1,4 +1,5 @@
 export const kMapTileType = "Map"
+export const kV2MapType = "map"
 export const kMapTileClass = "codap-map"
 export const kMapModelName = "MapModel"
