+++ v3/src/components/slider/slider-defs.ts
@@ -1,2 +1,3 @@
 export const kSliderTileType = "CodapSlider"
+export const kV2SliderType = "slider"
 export const kSliderTileClass = "codap-slider"
