+++ v3/src/components/slider/slider-registration.ts
@@ -1,21 +1,26 @@
 import { SetRequired } from "type-fest"
+import { V2Slider } from "../../data-interactive/data-interactive-component-types"
+import { registerComponentHandler } from "../../data-interactive/handlers/component-handler"
+import { errorResult } from "../../data-interactive/handlers/di-results"
+import { appState } from "../../models/app-state"
+import { GlobalValueManager } from "../../models/global/global-value-manager"
 import { registerTileComponentInfo } from "../../models/tiles/tile-component-info"
 import { ITileLikeModel, registerTileContentInfo } from "../../models/tiles/tile-content-info"
 import { getGlobalValueManager } from "../../models/tiles/tile-environment"
 import { ITileModelSnapshotIn } from "../../models/tiles/tile-model"
 import { toV3GlobalId, toV3Id } from "../../utilities/codap-utils"
+import { isAliveSafe } from "../../utilities/mst-utils"
+import { t } from "../../utilities/translation/translate"
 import { registerV2TileImporter } from "../../v2/codap-v2-tile-importers"
 import { isV2SliderComponent } from "../../v2/codap-v2-types"
 import { SliderComponent } from "./slider-component"
 import { SliderInspector } from "./slider-inspector"
-import { kSliderTileType, kSliderTileClass } from "./slider-defs"
+import { kSliderTileType, kSliderTileClass, kV2SliderType } from "./slider-defs"
 import { ISliderSnapshot, SliderModel, isSliderModel } from "./slider-model"
 import { SliderTitleBar } from "./slider-title-bar"
 import { AnimationDirections, AnimationModes, kDefaultAnimationDirection, kDefaultAnimationMode } from "./slider-types"
 import SliderIcon from '../../assets/icons/icon-slider.svg'
 import { kDefaultSliderName, kDefaultSliderValue } from "./slider-utils"
-import { t } from "../../utilities/translation/translate"
-import { isAliveSafe } from "../../utilities/mst-utils"
 
 export const kSliderIdPrefix = "SLID"
 
@@ -105,4 +110,61 @@ registerV2TileImporter("DG.SliderView", ({ v2Component, v2Document, sharedModelM
   }
 
   return sliderTile
+})
+
+registerComponentHandler(kV2SliderType, {
+  create({ values }) {
+    const {
+      animationDirection: _animationDirection, animationMode: _animationMode, globalValueName,
+      lowerBound, upperBound
+    } = values as V2Slider
+
+    if (globalValueName) {
+      const { document } = appState
+      const globalManager = document.content?.getFirstSharedModelByType(GlobalValueManager)
+      const global = globalManager?.getValueByName(globalValueName)
+      if (!global) {
+        return errorResult(t("V3.DI.Error.globalNotFound", { vars: [globalValueName] }))
+      }
+
+      // Multiple sliders for one global value are not allowed
+      let existingTile = false
+      document.content?.tileMap.forEach(sliderTile => {
+        if (isSliderModel(sliderTile.content) && sliderTile.content.globalValue.id === global.id) {
+          existingTile = true
+        }
+      })
+      if (existingTile) {
+        return errorResult(t("V3.DI.Error.noMultipleSliders", { vars: [globalValueName] }))
+      }
+
+      const animationDirection = _animationDirection != null
+        ? AnimationDirections[Number(_animationDirection)] : undefined
+      const animationMode = _animationMode != null ? AnimationModes[_animationMode] : undefined
+      return {
+        type: kSliderTileType,
+        animationDirection,
+        animationMode,
+        axis: { min: lowerBound, max: upperBound, place: "bottom" },
+        globalValue: global.id
+      } as SetRequired<ISliderSnapshot, "type">
+    }
+    // fallback to creating a default slider
+    return { type: kSliderTileType }
+  },
+  get(content) {
+    if (isSliderModel(content)) {
+      const animationDirection = AnimationDirections.findIndex(value => value === content.animationDirection)
+      const animationMode = AnimationModes.findIndex(value => value === content.animationMode)
+      return {
+        type: kV2SliderType,
+        animationDirection,
+        animationMode,
+        globalValueName: content.globalValue.name,
+        lowerBound: content.axis.min,
+        upperBound: content.axis.max,
+        value: content.globalValue.value
+      } as V2Slider
+    }
+  }
 })
