+++ v3/src/data-interactive/data-interactive-component-types.ts
@@ -1,22 +1,13 @@
-import { kCalculatorTileType } from "../components/calculator/calculator-defs"
-import { kCaseCardTileType } from "../components/case-card/case-card-defs"
-import { kCaseTableTileType } from "../components/case-table/case-table-defs"
-import { kGraphTileType } from "../components/graph/graph-defs"
-import { kMapTileType } from "../components/map/map-defs"
-import { kSliderTileType } from "../components/slider/slider-defs"
-import { kWebViewTileType } from "../components/web-view/web-view-defs"
+import { kCalculatorTileType, kV2CalculatorType } from "../components/calculator/calculator-defs"
+import { kCaseCardTileType, kV2CaseCardType } from "../components/case-card/case-card-defs"
+import { kCaseTableTileType, kV2CaseTableType } from "../components/case-table/case-table-defs"
+import { kGraphTileType, kV2GraphType } from "../components/graph/graph-defs"
+import { kMapTileType, kV2MapType } from "../components/map/map-defs"
+import { kSliderTileType, kV2SliderType } from "../components/slider/slider-defs"
+import { kV2GameType, kV2WebViewType, kWebViewTileType } from "../components/web-view/web-view-defs"
 
-export const kV2CalculatorType = "calculator"
-export const kV2CaseTableType = "caseTable"
-export const kV2CaseCardType = "caseCard"
-export const kV2GameType = "game"
-export const kV2GraphType = "graph"
-export const kV2GuideViewType = "guideView"
-export const kV2ImageType = "image"
-export const kV2MapType = "map"
-export const kV2SliderType = "slider"
-export const kV2TextType = "text"
-export const kV2WebViewType = "webView"
+// export const kV2ImageType = "image"
+// export const kV2TextType = "text"
 
 export const kComponentTypeV3ToV2Map: Record<string, string> = {
   [kCalculatorTileType]: kV2CalculatorType,
