+++ v3/src/data-interactive/handlers/component-handler-web-view.test.ts
@@ -1,8 +1,9 @@
+import { kV2GameType } from "../../components/web-view/web-view-defs"
 import { IWebViewModel, isWebViewModel } from "../../components/web-view/web-view-model"
 import { kWebViewIdPrefix } from "../../components/web-view/web-view-registration"
 import { appState } from "../../models/app-state"
 import { toV3Id } from "../../utilities/codap-utils"
-import { kV2GameType, V2Game, V2WebView } from "../data-interactive-component-types"
+import { V2Game, V2WebView } from "../data-interactive-component-types"
 import { DIComponentInfo } from "../data-interactive-types"
 import { diComponentHandler } from "./component-handler"
 import { testGetComponent } from "./component-handler-test-utils"
