+++ v3/src/models/shared/shared-data-utils.ts
@@ -1,4 +1,5 @@
 import { IAnyStateTreeNode } from "@concord-consortium/mobx-state-tree"
+import { toV2Id } from "../../utilities/codap-utils"
 import { IDataSet } from "../data/data-set"
 import { ITileContentModel } from "../tiles/tile-content"
 import { ITileModel } from "../tiles/tile-model"
@@ -11,6 +12,12 @@ import { ISharedDataSet, isSharedDataSet, kSharedDataSetType, SharedDataSet } fr
 export function getSharedDataSets(node: IAnyStateTreeNode): ISharedDataSet[] {
   const sharedModelManager = getSharedModelManager(node)
   return sharedModelManager?.getSharedModelsByType<typeof SharedDataSet>(kSharedDataSetType) ?? []
+}
+
+export function getDataSetByNameOrId(node: IAnyStateTreeNode, nameOrId?: string): Maybe<IDataSet> {
+  return getSharedDataSets(node).find(({ dataSet }) => {
+    return dataSet.id === nameOrId || dataSet.name === nameOrId || `${toV2Id(dataSet.id)}` === nameOrId
+  })?.dataSet
 }
 
 export function getSharedDataSetFromDataSetId(node: IAnyStateTreeNode, id: string): ISharedDataSet | undefined {
