+++ v3/src/models/tiles/tile-content-info.ts
@@ -34,6 +34,7 @@ export interface ITileContentInfo {
   defaultName?: (options?: IDefaultContentOptions) => string | undefined
   titleBase?: string;
   getTitle: (tile: ITileLikeModel) => string | undefined;
+  getV2Type?: (content: ITileContentModel) => string;
   metadataClass?: typeof TileMetadataModel;
   isSingleton?: boolean; // Only one instance of a tile is open per document (calculator and guide)
   hideOnClose?: boolean;
